Création classe Intervenants pour gérer la liste des intervenants Mise en place d'un id de mission et d'un id d'intervenant Remplacement de int par size_t Calcul de la solution initial d'un chromosome

+++ src/main.cpp
@@ -24,10 +24,10 @@ int main(int argc, char **argv)
 
 	// valeurs par defaut
 	int nb_generation = 50;
-	int taille_population = 20;
+	size_t taille_population = 20;
 	float taux_croisement = 0.8;
 	float taux_mutation = 0.5;
-	int taille_chromosome = 100;
+	size_t taille_chromosome = 100;
 	string folderInstance = "instances/100-10";
 
 	if (argc == 7)
